Remove references to snapshots and make tests pass
