+++ custom/opm/tests/case_reports.py
@@ -101,7 +101,6 @@ class MockCaseRow(OPMCaseRow):
     def __init__(self, case, report, data_provider=None, child_index=1):
         self.case = case
         self.report = report
-        self.report.snapshot = None
         self.report.is_rendered_as_email = None
         self.report._data_provider = data_provider or MockDataProvider(report.datespan.enddate.date())
         super(MockCaseRow, self).__init__(case, report, child_index=child_index)
